requirements.txt

torch==2.0.0             
torchvision==0.15.1      
torchaudio==2.0.1        
torch-geometric         
torch-scatter           
torch-sparse            
torch-cluster           
rdkit                   

pytorch-lightning>=2.0,<2.1
torchmetrics>=0.11
transformers>=4.20
peft>=0.4


numpy>=1.21
scipy>=1.8
pandas>=1.4
scikit-learn>=1.1
PyYAML>=6.0

tqdm>=4.60
joblib>=1.1
requests>=2.20
lmdb>=1.3
biopython>=1.79
omegaconf>=2.1 
hydra-core>=1.1 

pomegranate>=0.14,<1.0
moses-lib>=1.2
fcd-torch>=1.1


matplotlib>=3.5
seaborn>=0.11
tensorboard
wandb

